requirements_clean.txt
Flask==3.1.0
Flask-Mail==0.10.0
Flask-WTF==1.2.2
WTForms==3.2.1
python-dotenv==1.0.0
email-validator==2.2.0
gunicorn==21.2.0